+++ lib/local_image_provider.dart
@@ -18,13 +18,27 @@ class LocalImageProvider {
   static final LocalImageProvider _instance =
       LocalImageProvider.withMethodChannel(lipChannel);
   final MethodChannel channel;
+  bool _initWorked = false;
   factory LocalImageProvider() => _instance;
   @visibleForTesting
   LocalImageProvider.withMethodChannel(this.channel);
 
+  /// True if [initialize] succeeded
+  bool get isAvailable => _initWorked;
+
+  Future<bool> initialize() async {
+    if (_initWorked) {
+      return Future.value(_initWorked);
+    }
+    _initWorked = await channel.invokeMethod('initialize');
+    return _initWorked;
+  }
+
   /// Returns the list of [LocalAlbum] available on the device matching the [localAlbumType]
-  Future<List<LocalAlbum>> getAlbums(
-      LocalAlbumType localAlbumType) async {
+  Future<List<LocalAlbum>> getAlbums(LocalAlbumType localAlbumType) async {
+    if (!_initWorked) {
+      throw LocalImageProviderNotInitializedException();
+    }
     final List<dynamic> albums =
         await channel.invokeMethod('albums', localAlbumType.value);
     return albums.map((albumJson) {
@@ -39,10 +53,13 @@ class LocalImageProvider {
   /// This list may be empty if there are no photos on the device or the
   /// user has denied permission to see their local photos.
   Future<List<LocalImage>> getLatest(int maxImages) async {
+    if (!_initWorked) {
+      throw LocalImageProviderNotInitializedException();
+    }
     final List<dynamic> photos =
         await channel.invokeMethod('latest_images', maxImages);
     return photos.map((photoJson) {
-      print(photoJson);
+      // print(photoJson);
       Map<String, dynamic> photoMap = jsonDecode(photoJson);
       return LocalImage.fromJson(photoMap);
     }).toList();
@@ -54,8 +71,15 @@ class LocalImageProvider {
   /// The returned image will maintain its aspect ratio while fitting within the given dimensions
   /// [height], [width]. The [id] to use is available from a returned LocalImage.
   Future<Uint8List> imageBytes(String id, int height, int width) async {
+    if (!_initWorked) {
+      throw LocalImageProviderNotInitializedException();
+    }
     final Uint8List photoBytes = await channel.invokeMethod(
         'image_bytes', {'id': id, 'pixelHeight': height, 'pixelWidth': width});
     return photoBytes;
   }
 }
+
+/// Thrown when a method is called that requires successful
+/// initialization first. See [initialize]
+class LocalImageProviderNotInitializedException implements Exception {}