+++ test/local_album_test.dart
@@ -2,10 +2,10 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:local_image_provider/local_album.dart';
 
 void main() {
-    test('Correct type returned', () async {
-      expect( LocalAlbumType.fromInt(0), LocalAlbumType.all );
-      expect( LocalAlbumType.fromInt(1), LocalAlbumType.user );
-      expect( LocalAlbumType.fromInt(2), LocalAlbumType.generated );
-      expect( LocalAlbumType.fromInt(3), null );
-    });
+  test('Correct type returned', () async {
+    expect(LocalAlbumType.fromInt(0), LocalAlbumType.all);
+    expect(LocalAlbumType.fromInt(1), LocalAlbumType.user);
+    expect(LocalAlbumType.fromInt(2), LocalAlbumType.generated);
+    expect(LocalAlbumType.fromInt(3), null);
+  });
 }