+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+
+## 0.5.2
+### Breaking
+* Added initialize method, now required before using the plugin
+* Methods are no longer static, you must now create an instance and use that
+  
+### Updates
+* Added permission checking to both iOS and Android
+ 
 ## 0.5.1
 * Fixes issues pointed out by pub_dev analysis
 * updated description in pubspec